requirements.txt
wget
allennlp==0.9.0
conda==4.3.16
Keras==2.1.5
numpy==1.18.1
pandas==0.23.4
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
scikit-learn==0.20.2
scipy==1.1.0
tensorflow==1.13.1

